% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{activate}
\alias{activate}
\title{Activate orderly for current reposotory}
\usage{
activate(verbose = NULL)
}
\value{
Nothing.
}
\description{
Activate orderly for the current repository.  Run this from
anywhere within an orderly archive (i.e., a path containing
orderly_config.yml; or where one of its parent directories do). We
check what the required orderly minimum version is and set up the
orderly namespace for the appropriate version.
}
\author{
Richard Fitzjohn
}
