% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{auto}
\alias{auto}
\title{Automatically configure orderly}
\usage{
auto()
}
\value{
Nothing, called for side effects only
}
\description{
Automatically set up the orderly namespace depending on the
context in which this function is called.
}
\details{
If called from within an orderly repository, it will activate the
required version using \code{\link[=activate]{activate()}} otherwise it will use the
globally preferred version with \code{\link[=use]{use()}}
}
