% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{use}
\alias{use}
\title{Set orderly version}
\usage{
use(version = NULL, verbose = NULL)
}
\arguments{
\item{version}{Either \code{NULL} (in which case we use global
preferences) or a number \code{1} or \code{2}.}
}
\value{
Nothing, called for side effects only.
}
\description{
Set up orderly based on your global preferences. This will look at
the R option (\code{orderly.version}) and the environment variable
\code{ORDERLY_VERSION}, in that order, falling back on the second
version. Or pass in a version explicitly.
}
